requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argo-ai==0.4.1
attrs==25.3.0
beautifulsoup4==4.13.5
beaver-db==0.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
cryptography==46.0.1
defusedxml==0.7.1
distro==1.9.0
fastembed==0.7.3
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
loguru==0.7.3
lxml==6.0.1
magika==0.6.2
mammoth==1.10.0
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.5.0
numpy==2.3.3
onnxruntime==1.22.1
openai==1.107.3
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pillow==11.3.0
protobuf==6.32.1
py_rust_stemmers==0.1.5
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
